pyproject.toml
[project]
name = "quoin-api"
version = "0.6.0"
description = "The architectural cornerstone for high-performance, scalable Python services"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Balakumaran Manoharan", email = "bala@balakmran.dev" }
]
license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    "fastapi[standard]==0.135.3",
    "sqlmodel==0.0.38",
    "psycopg[binary]==3.3.3",
    "structlog>=25.5.0",
    "pydantic-settings>=2.13.1",
    "alembic>=1.18.4",
    "greenlet==3.4.0",
    "asyncpg>=0.31.0",
    "jinja2>=3.1.6",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "opentelemetry-exporter-otlp>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "PyJWT[crypto]>=2.12.1",
    "httpx>=0.28.1",
]

[project.urls]
Repository = "https://github.com/balakmran/quoin-api"
Documentation = "https://balakmran.github.io/quoin-api/"

[dependency-groups]
dev = [
    { include-group = "tooling" },
    { include-group = "test" },
    { include-group = "docs" },
]
tooling = [
    "ruff==0.15.11",
    "ty==0.0.31",
    "prek==0.3.8",
]
test = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
docs = [
    "zensical==0.0.32",
    "mkdocstrings-python>=2.0.3",
]

[tool.ruff]
line-length = 80
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF", "D"]
ignore = ["D100", "D104"] # Ignore missing docstring in public module/package
pydocstyle.convention = "google"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.params.Depends",
    "fastapi.Query",
    "fastapi.Path",
    "fastapi.Body",
    "fastapi.Security",
    "fastapi.Header",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["app"]
branch = true
concurrency = ["thread", "greenlet"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
